+++ src/django/users/tests/test_models.py
@@ -8,6 +8,7 @@ from django.utils import timezone
 from users.models import PlanItLocation, PlanItOrganization, PlanItUser
 from planit_data.models import GeoRegion, OrganizationWeatherEvent, WeatherEvent, WeatherEventRank
 from planit_data.tests.factories import GeoRegionFactory
+from users.tests.factories import OrganizationFactory, UserFactory
 
 
 class OrganizationTestCase(TestCase):
@@ -273,7 +274,7 @@ class LocationManagerTestCase(TestCase):
 
 class PlanItUserTestCase(TestCase):
     def test_default_user_model(self):
-        """Ensure PlanItUser is the default user models
+        """Ensure PlanItUser is the default user model.
 
         `./manage.py createsuperuser` uses get_user_model to determine what class to invoke
         create_superuser on, this ensures the command will be directed to the correct class.
@@ -313,3 +314,66 @@ class PlanItUserTestCase(TestCase):
         self.assertEqual(0, user.organizations.all().count(), 'User should have no organizations')
         self.assertEqual(user.primary_organization, None,
                          'User should have no primary organization')
+
+    def test_add_user_via_email_no_org(self):
+        """Ensure that adding a user via email properly sets the organization."""
+        org = OrganizationFactory()
+        existing_user_no_org_email = 'existinguser@local.gov'
+        user = UserFactory(email=existing_user_no_org_email)
+        # The factory constructs the user with a new Organization by default, which is normally okay
+        # but in this case we don't want that, so clear it out.
+        user.organizations.clear()
+        user.primary_organization = None
+        user.save()
+        # Adding a user that has no org should update their primary_organization
+        self.assertIsNone(user.primary_organization)
+        new_user, created, added = PlanItUser.objects.add_via_email_to_organization(
+            existing_user_no_org_email, org
+        )
+        self.assertEqual(new_user, user)
+        self.assertFalse(created)
+        self.assertTrue(added)
+        self.assertEqual(new_user.primary_organization, org)
+        self.assertIn(org, new_user.organizations.all())
+
+    def test_add_user_via_email_same_org(self):
+        """Ensure that adding a user that has an org to that same org does nothing."""
+        org = OrganizationFactory()
+        existing_user_org_email = 'existinguser+org1@local.gov'
+        user = UserFactory(email=existing_user_org_email, primary_organization=org)
+        new_user, created, added = PlanItUser.objects.add_via_email_to_organization(
+            existing_user_org_email, org
+        )
+        self.assertEqual(new_user, user)
+        self.assertFalse(created)
+        self.assertFalse(added)
+
+    def test_add_user_via_email_different_org(self):
+        """Ensure that adding a user that has an org adds the org to that user's organizations."""
+        org1 = OrganizationFactory()
+        org2 = OrganizationFactory()
+        existing_user_org1_email = 'existinguser+org1@local.gov'
+        user = UserFactory(email=existing_user_org1_email, primary_organization=org1)
+        self.assertEqual(user.primary_organization, org1)
+        new_user, created, added = PlanItUser.objects.add_via_email_to_organization(
+            existing_user_org1_email, org2
+        )
+        self.assertEqual(new_user, user)
+        self.assertEqual(new_user.primary_organization, org1)
+        self.assertFalse(created)
+        self.assertTrue(added)
+        self.assertIn(org2, new_user.organizations.all())
+
+    def test_add_user_via_email_new_user(self):
+        """Ensure that adding a user that doesn't exist creates them in the correct organization."""
+        org = OrganizationFactory()
+        new_user_email = 'newuser@local.gov'
+        self.assertEqual(PlanItUser.objects.all().count(), 0)
+        new_user, created, added = PlanItUser.objects.add_via_email_to_organization(
+            new_user_email, org
+        )
+        self.assertEqual(PlanItUser.objects.all().count(), 1)
+        self.assertTrue(created)
+        self.assertFalse(added)
+        self.assertEqual(new_user.primary_organization, org)
+        self.assertIn(org, new_user.organizations.all())